+++ Packages/cn.etetet.login/Scripts/Hotfix/Client/Login/ClientSenderComponentSystem.cs
@@ -25,7 +25,7 @@ namespace ET.Client
                 return;
             }
 
-            int fiberId = self.fiberId;
+            long fiberId = self.fiberId;
             self.fiberId = 0;
             await self.Fiber().RemoveFiber(fiberId);
         }
