+++ Packages/cn.etetet.login/Scripts/Hotfix/Client/NetClient/NetComponentOnReadInvoker_NetClient.cs
@@ -27,7 +27,7 @@ namespace ET.Client
                 case IMessage iActorMessage:
                 {
                     // 扔到Main纤程队列中
-                    int parentFiberId = fiber.Root.GetComponent<FiberParentComponent>().ParentFiberId;
+                    long parentFiberId = fiber.Root.GetComponent<FiberParentComponent>().ParentFiberId;
                     fiber.Root.GetComponent<ProcessInnerSender>().Send(new FiberInstanceId(parentFiberId), iActorMessage);
                     break;
                 }
